% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePWProfiles.R
\name{CalculatePWProfiles}
\alias{CalculatePWProfiles}
\title{Calculate pathway enrichment profile}
\usage{
CalculatePWProfiles(object, gmt_path, method = "gsva", min.sz = 5,
  max.sz = 200, parallel.sz = NULL)
}
\arguments{
\item{object}{A Tempora object containing a gene expression matrix and metadata (cluster identity and )}

\item{gmt_path}{Local path to database of pathways or genesets organized as a .gmt file. Genesets files in GMT format can be downloaded at http://baderlab.org/GeneSets. Please ensure}

\item{method}{Method used to estimate pathway enrichment profile per cluster. Can be "gsva", "ssgsea", "zscore" or "plage", default to "gsva". See ?gsva for more information.}

\item{min.sz}{Minimum size of the genesets used in enrichment estimation, set to 5 genes by default.}

\item{max.sz}{Maximum size of the genesets used in enrichment estimation, set to 200 genes by default.}

\item{parallel.sz}{Type of cluster architecture when using \code{snow}. If NULL, no parallelization will be used.}
}
\value{
An updated Tempora object containing the pathway enrichment profiles of each cluster, which can be accessed at \code{object@cluster.pathways}
CalculatePWProfiles
}
\description{
Calculate cluster average gene expression profile and determine the pathway enrichment profile of each cluster
}
