% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporaObjClass.R
\name{CreateTemporaObject}
\alias{CreateTemporaObject}
\title{Create new Tempora object from a processed gene expression matrix}
\usage{
CreateTemporaObject(
  exprMatrix,
  meta.data,
  timepoint_order,
  cluster_labels = NULL,
  cell_markers = NULL
)
}
\arguments{
\item{exprMatrix}{A normalized gene expression matrix containing cells from all timepoints of the time-series study. Batch effect correction is highly recommended before normalization.}

\item{meta.data}{A dataframe of meta data for all cells in the expression matrix. Each column is a feature and each row stores single cell information. At minimum, this dataframe should contain two columns: a "Clusters" column storing the clustering identity and a "Timepoints" column storing the timepoint when each cell comes from}

\item{timepoint_order}{An ordered vector of timepoint names from early to late}

\item{cluster_labels}{(Optional) A vector of cluster annotations (cell types, cell states, cell cycles, etc.), ordered alphanumerically by cluster names. If NULL and \code{cell_markers} is given, automatic cluster annotation using GSVA will be performed. If both are NULL, cluster numbers will be used to label the trajectory plot.}

\item{cell_markers}{(Optional) A list of possible cell types found in the dataset and their marker genes to be used for automatic cell type identification. If NULL and no \code{cluster_labels} is given, cluster numbers will be used to label the trajectory plot.}
}
\value{
A Tempora object containing the expression matrix and metadata
}
\description{
Create new Tempora object from a processed gene expression matrix
}
\examples{
\dontrun{tempora_dara <- CreateTemporaObject(exprMatrix, meta.data)}

}
