% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyVaryingPWs.R
\name{IdentifyVaryingPWs}
\alias{IdentifyVaryingPWs}
\title{Calculate pathway enrichment profile}
\usage{
IdentifyVaryingPWs(object, pval_threshold = 0.05)
}
\arguments{
\item{object}{A Tempora object containing a gene expression matrix and metadata (cluster identity and )}

\item{pathwaygmt}{A database of pathways oirganized as a .gmt file}

\item{method}{Method used to calculate pathway enrichment profile. Can be "gsva", "ssgsea", "zscore" or "plage". See ?gsva for more information.}
}
\description{
Calculate cluster average gene expression profile and determine the pathway enrichment profile of each cluster
}
