% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyVaryingPWsParallel.R
\name{IdentifyVaryingPWsParallel}
\alias{IdentifyVaryingPWsParallel}
\title{Calculate temporally changing pathways (parallel version)}
\usage{
IdentifyVaryingPWsParallel(object, pval_threshold = 0.05)
}
\arguments{
\item{object}{A Tempora object}

\item{pval_threshold}{P-value threshold to determine the significance of pathway enrichment over time. Default to 0.05.}
}
\description{
Identify the pathways that change over time by fitting a generalized additive model
}
\examples{
\dontrun{tempora_data <- IdentifyVaryingPWsParallel(tempora_data, pval_threshold = 0.05)}
}
