% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{PlotTrajectory}
\alias{PlotTrajectory}
\title{Visualize the trajectory}
\usage{
PlotTrajectory(object, layout = NULL, ...)
}
\arguments{
\item{object}{A Tempora object}

\item{layout}{A 2-column matrix containing the x- and y-coordinates of all vertices in the graph. If NULL, the layout will be obtained using Sugiyama layout}

\item{...}{Any additional arguments to plot.igraph}
}
\description{
Reduce the dimensionality of the pathway enrichment matrix and build the information-based cluster-cluster network
}
\examples{
\dontrun{tempora_data <- PlotTrajectory(tempora_data)}
}
