% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotVaryingPWs.R
\name{PlotVaryingPWs}
\alias{PlotVaryingPWs}
\title{Plot temporally changing pathways}
\usage{
PlotVaryingPWs(object)
}
\arguments{
\item{object}{A Tempora object}
}
\description{
Plot the expression of temporally changing pathways as identified by IdentifyVaryingPWs()
}
\examples{
\dontrun{tempora_data <- IdentifyVaryingPWs(tempora_data, pval_threshold = 0.05)}
}
