% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyCellTypes.R
\name{IdentifyCellTypes}
\alias{IdentifyCellTypes}
\title{Calculate pathway enrichment profile}
\usage{
IdentifyCellTypes(exprMatrix, cluster_labels, cell_markers)
}
\arguments{
\item{exprMatrix}{A gene expression matrix, with genes in rows and cells in columns.}

\item{cluster_labels}{A named vector of cluster identifier for each cell in the gene expression matrix}

\item{cell_markers}{A list of possible cell types found in the dataset and their marker genes.}
}
\value{
A vector of cell types inferred from the expression of marker genes provided
CalculatePWProfiles
}
\description{
Calculate cluster average gene expression profile and determine the pathway enrichment profile of each cluster
}
