% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyVaryingPWs.R
\name{IdentifyVaryingPWs}
\alias{IdentifyVaryingPWs}
\title{Calculate pathway enrichment profile}
\usage{
IdentifyVaryingPWs(object, pval_threshold = 0.05)
}
\arguments{
\item{object}{A Tempora object}

\item{pval_threshold}{P-value threshold to determine the significance of pathway enrichment over time. Default to 0.05.}
}
\description{
Calculate cluster average gene expression profile and determine the pathway enrichment profile of each cluster
}
\examples{
\dontrun{tempora_data <- IdentifyVaryingPWs(tempora_data, pval_threshold = 0.05)}
}
