% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporaObjClass.R
\name{ImportSeuratObject}
\alias{ImportSeuratObject}
\title{Import data from a Seurat object}
\usage{
ImportSeuratObject(
  seuratobj,
  assayType = "",
  assaySlot = NA,
  clusters,
  timepoints,
  timepoint_order,
  cluster_labels = NULL,
  cell_markers = NULL
)
}
\arguments{
\item{seuratobj}{A Seurat or SingleCellExperiment object containing the normalized gene expression matrix and clustering result}

\item{assayType}{A length-one character vector representing the assay object
in which the expression data is stored in the input object. For Seurat v1
or v2 objects, set this to "". For Seurat v3 objects, this is often "RNA".
For SingleCellExperiment objects, this is often "logcounts".}

\item{assaySlot}{An optional length-one character vector representing the
slot of the Seurat v3 \code{\link[Seurat]{Assay}} object to use. In Seurat
v3, normalized data is stored in the "data" slot, and counts in the
"counts" slot.}

\item{clusters}{Name of the column in the meta.data dataframe containing the cluster identity of all cells in the dataset}

\item{timepoints}{Name of the column in the meta.data dataframe containing the collection time of all cells in the dataset}

\item{timepoint_order}{An ordered vector of timepoint names from early to late}

\item{cluster_labels}{(Optional) A vector of cluster annotations (cell types, cell states, cell cycles, etc.), ordered alphanumerically by cluster names. If NULL and \code{cell_markers} is given, automatic cluster annotation using GSVA will be performed. If both are NULL, cluster numbers will be used to label the trajectory plot.}

\item{cell_markers}{(Optional) A list of possible cell types found in the dataset and their marker genes to be used for automatic cell type identification. If NULL and no \code{cluster_labels} is given, cluster numbers will be used to label the trajectory plot.}
}
\value{
A Tempora object containing the expression matrix and metadata
}
\description{
Imports gene expression matrix and other metadata from a seurat object
}
\examples{
\dontrun{tempora_data <- ImportSeuratObject(seurat_object, assayType="", clusters = "res.0.3", timepoints = "collection_time",
timepoint_order = c("0H", "24H", "48H", "72H"), cluster_labels = c("Stem cells", "Differentiated cells"))}
}
