% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporaObjClass.R
\name{ImportSeuratObject}
\alias{ImportSeuratObject}
\title{Import data from a Seurat object}
\usage{
ImportSeuratObject(seuratobj, clusters, timepoints, timepoint_order,
  cluster_labels)
}
\arguments{
\item{seuratobj}{A Seurat object containing the normalized gene expression matrix and clustering result}

\item{clusters}{Name of the column in the meta.data dataframe containing the cluster identity of all cells in the dataset}

\item{timepoints}{Name of the column in the meta.data dataframe containing the collection time of all cells in the dataset}

\item{timepoint_order}{An ordered vector of timepoint names from early to late}

\item{cluster_labels}{A vector of cluster annotations (cell types, cell states, cell cycles, etc.), ordered alphanumerically by cluster names. If NULL, cluster numbers will be used to label the trajectory plot}
}
\value{
A Tempora object containing the expression matrix and metadata
}
\description{
Imports gene expression matrix and other metadata from a seurat object
}
\examples{
tempora_data <- ImportSeuratObject(seurat_object, clusters = "res.0.3", timepoints = "collection_time", timepoint_order = c("0H", "24H", "48H", "72H"), cluster_labels = c("Stem cells", "Differentiated cells"))
Import data from a Seurat object
}
