% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.R
\name{PlotTrajectory}
\alias{PlotTrajectory}
\title{Visualize the trajectory}
\usage{
PlotTrajectory(object)
}
\arguments{
\item{object}{A Tempora object}
}
\description{
Reduce the dimensionality of the pathway enrichment matrix and build the information-based cluster-cluster network
}
\examples{
\dontrun{tempora_data <- PlotTrajectory(tempora_data)}
}
